Fixed bug in dist-upgrade.yml
There's no .stat property for dist_upgrade_register_etc_services if the conditions in the previous statement ('Check if /etc/services.d exists') were not met. In that case, the variable is a dict with a message stating that the condition was not met and a changed = false. So the correct condition for 'Assemble /etc/services' would be to look if stat was run at all.

I checked and this issue should only persist in this case and nowhere else in this playbook.