Fixed bug in dist-upgrade.yml when dist_upgrade_register_upgrade is undefined or unchanged
In cases where dist_upgrade_register_upgrade is undefined or unchanged, the 'Check if /etc/services' will register a dict without any of the stat return parameters. This leads to 'Assemble /etc/services' to fail with '"dict object" has no attribute "stat"'.

Adding another check in the when statement to see if the previous task was skipped, we will not encounter this error as it does skip there already.